Restore naming of GMM file, fix package cycle
