+++ subprojects/gmm-publish/src/main/java/org/gradle/api/publish/gmm/internal/artifact/DefaultGMMArtifact.java
@@ -20,7 +20,7 @@ import com.google.common.collect.ImmutableSet;
 import org.gradle.api.Incubating;
 import org.gradle.api.internal.tasks.DefaultTaskDependency;
 import org.gradle.api.internal.tasks.TaskDependencyFactory;
-import org.gradle.api.publish.gmm.GMMArtifact;
+import org.gradle.api.publish.gmm.artifact.GMMArtifact;
 import org.gradle.api.publish.tasks.GenerateModuleMetadata;
 import org.gradle.api.tasks.TaskDependency;
