+++ subprojects/gmm-publish/src/main/java/org/gradle/api/publish/gmm/internal/publication/DefaultGMMPublication.java
@@ -29,7 +29,8 @@ import org.gradle.api.internal.attributes.ImmutableAttributes;
 import org.gradle.api.internal.component.SoftwareComponentInternal;
 import org.gradle.api.internal.tasks.TaskDependencyFactory;
 import org.gradle.api.publish.Publication;
-import org.gradle.api.publish.gmm.GMMArtifact;
+import org.gradle.api.publish.gmm.artifact.GMMArtifact;
+import org.gradle.api.publish.gmm.internal.GMMProjectIdentity;
 import org.gradle.api.publish.internal.PublicationArtifactSet;
 import org.gradle.api.publish.internal.versionmapping.VersionMappingStrategyInternal;
 import org.gradle.api.tasks.TaskProvider;
